+++ unstructured/chunking/base.py
@@ -33,6 +33,20 @@ from unstructured.documents.elements import (
 )
 from unstructured.utils import lazyproperty
 
+# -- CONSTANTS -----------------------------------
+
+CHUNK_MAX_CHARS_DEFAULT: int = 500
+"""Hard-max chunk-length when no explicit value specified in `max_characters` argument."""
+
+CHUNK_MULTI_PAGE_DEFAULT: bool = True
+"""When False, respect page-boundaries (no two elements from different page in same chunk).
+
+Only operative for "by_title" chunking strategy.
+w"""
+
+
+# -- TYPES ---------------------------------------
+
 BoundaryPredicate: TypeAlias = Callable[[Element], bool]
 """Detects when element represents crossing a semantic boundary like section or page."""
 
@@ -94,8 +108,8 @@ class ChunkingOptions:
     def __init__(
         self,
         combine_text_under_n_chars: Optional[int] = None,
-        max_characters: int = 500,
-        multipage_sections: bool = True,
+        max_characters: int = CHUNK_MAX_CHARS_DEFAULT,
+        multipage_sections: bool = CHUNK_MULTI_PAGE_DEFAULT,
         new_after_n_chars: Optional[int] = None,
         overlap: int = 0,
         overlap_all: bool = False,
@@ -113,8 +127,8 @@ class ChunkingOptions:
     def new(
         cls,
         combine_text_under_n_chars: Optional[int] = None,
-        max_characters: int = 500,
-        multipage_sections: bool = True,
+        max_characters: int = CHUNK_MAX_CHARS_DEFAULT,
+        multipage_sections: bool = CHUNK_MULTI_PAGE_DEFAULT,
         new_after_n_chars: Optional[int] = None,
         overlap: int = 0,
         overlap_all: bool = False,